+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.3.0
+
+- jumped to 1.3.0 to make it consistent with result_flow package
+- added stack trace to all network errors and bad response parser
+
 ## 1.0.2
 
 - Removed angle brackets from docs